Fixed TX signal configuration
Previously, when the SIGNAL CONFIGURATION button was pressed, the PLUTO was configured, but not the different parts of the signal. When the TRANSMIT SIGNAL button was pressed, the parts of the signal were configured and then sent, which generated a delay in sending the signal. Now, this has been corrected, and the parts of the signal are prepared when it is configured, to be sent without prior delay when the TRANSMIT SIGNAL button is pressed